+++ lib/db.bash
@@ -52,7 +52,7 @@ function odoo_db_list {
     execu python -c "\"$python_cmd\"";
 }
 
-# odoo_db_exists [odoo_conf_file]
+# odoo_db_exists <dbname> [odoo_conf_file]
 function odoo_db_exists {
     local db_name=$1;
     local conf_file=${2:-$ODOO_CONF_FILE};
@@ -69,6 +69,72 @@ function odoo_db_exists {
     fi
 }
 
+# odoo_db_dump <dbname> <file-path> [format [odoo_conf_file]]
+# dump database to specified path
+function odoo_db_dump {
+    local db_name=$1;
+    local db_dump_file=$2;
+    local conf_file=$ODOO_CONF_FILE;
+
+    if [ -f "$3" ]; then
+        conf_file=$3;
+    elif [ ! -z $3 ]; then
+        local format=$3;
+        local format_opt=", '$format'";
+
+        if [ -f "$4" ]; then
+            conf_file=$4;
+        fi
+    fi
+
+    local python_cmd="import erppeek; cl=erppeek.Client(['-c', '$conf_file']);";
+    python_cmd="$python_cmd dump=cl.db.dump(cl._server.tools.config['admin_passwd'], '$db_name' $format_opt).decode('base64');";
+    python_cmd="$python_cmd open('$db_dump_file', 'wb').write(dump);";
+    
+    if execu python -c "\"$python_cmd\""; then
+        echov "Database named '$db_name' dumped to '$db_dump_file'!";
+        return 0;
+    else
+        echo "Database '$db_name' fails on dump!";
+        return 1;
+    fi
+}
+
+
+# odoo_db_backup <dbname> [format [odoo_conf_file]]
+# if second argument is file and it exists, then it used as config filename
+# in other cases second argument is treated as format, and third (if passed) is treated as conf file
+function odoo_db_backup {
+    if [ -z $BACKUP_DIR ]; then
+        echo "Backup dir is not configured. Add 'BACKUP_DIR' variable to your 'odoo-helper.conf'!";
+        return 1;
+    fi
+
+    local FILE_SUFFIX=`date -I`.`random_string 4`;
+    local db_name=$1;
+    local db_dump_file="$BACKUP_DIR/db-backup-$db_name-$FILE_SUFFIX.backup";
+
+    odoo_db_dump $db_name $db_dump_file $2 $3;
+}
+
+# odoo_db_restore <dbname> <dump_file> [odoo_conf_file]
+function odoo_db_restore {
+    local db_name=$1;
+    local db_dump_file=$2;
+    local conf_file=${3:-$ODOO_CONF_FILE};
+
+    local python_cmd="import erppeek; cl=erppeek.Client(['-c', '$conf_file']);";
+    python_cmd="$python_cmd res=cl.db.restore(cl._server.tools.config['admin_passwd'], '$db_name', open('$db_dump_file', 'rb').read().encode('base64'));";
+    python_cmd="$python_cmd exit(0 if res else 1);";
+    
+    if execu python -c "\"$python_cmd\""; then
+        echov "Database named '$db_name' restored from '$db_dump_file'!";
+        return 0;
+    else
+        echov "Database '$db_name' fails on restore from '$db_dump_file'!";
+        return 1;
+    fi
+}
 
 # Command line args processing
 function odoo_db_command {
@@ -78,6 +144,9 @@ function odoo_db_command {
         $SCRIPT_NAME db exists <name> [odoo_conf_file]
         $SCRIPT_NAME db create <name> [odoo_conf_file]
         $SCRIPT_NAME db drop <name> [odoo_conf_file]
+        $SCRIPT_NAME db dump <name> <dump_file_path> [format [odoo_conf_file]]
+        $SCRIPT_NAME db backup <name> [format [odoo_conf_file]]
+        $SCRIPT_NAME db restore <name> <dump_file_path> [odoo_conf_file]
 
     ";
 
@@ -105,9 +174,23 @@ function odoo_db_command {
                 odoo_db_drop "$@";
                 exit;
             ;;
+            dump)
+                shift;
+                odoo_db_dump "$@";
+                exit;
+            ;;
+            backup)
+                shift;
+                odoo_db_backup "$@";
+                exit;
+            ;;
+            restore)
+                shift;
+                odoo_db_restore "$@";
+                exit;
+            ;;
             exists)
                 shift;
-                local VERBOSE=1;
                 odoo_db_exists "$@";
                 exit;
             ;;